Changed default resuseMtx to False
Honeybee[+] unlike Honeybee loads the results from a previous study if
available. This was designed to reuse the results of the studies instead
of re-running them. This behavior has made a lot of confusion for users.
For now all the reuse options are set to False by default until we
implement a smarter change detection to re-run the analysis based on
changes in inputs. You can still set them to True to reuse the results
if you need.